# swagger.json
{
    "swagger":"2.0",
    "info":{
        "version": "1.0.0",
        "title": "API Monitor",
        "description": "Flymonitor documentation",
        "licence":{
            "name":"MIT",
            "url":"https://flybyte.com.br"
        }
    },
    "paths":
...
        "/admin/registercam":{
            "post":
            {
                "tags":["Admin Calls"],
                "summary": "Add new camera",
                "description":"call responsible for adding a new camera in the admin and client",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                },
                {
                    "in": "body",
                    "name":"userData",
                    "description": false,
                    "required": true,
                    "properties":
                    {
                        "emailAdmin":{"type":"string"},
                        "nomeClient:":{"type":"string"},
                        "nomeCamera":{"type":"string"},
                        "descricao":{"type":"string"},
                        "RTSP":{"type":"string"}
                    }              
                }
            ]
        },
        "/admin/consultaCAM":{
            "post":
            {
                "tags":["Admin Calls"],
                "summary": "Consult a camera",
                "description":"Call responsible for consulting all cameras registered by the admin",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                },
                {
                    "in": "body",
                    "name":"userData",
                    "description": false,
                    "required": true,
                    "properties":
                    {
                        "email":{"type":"string"}
                    }              
                }
            ]
        },
        "/admin/listclients":{
            "post":
            {
                "tags":["Admin Calls"],
                "summary": "List client at admin",
                "description":"Call responsible for consulting all cameras registered by the admin",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                },
                {
                    "in": "body",
                    "name":"userData",
                    "description": false,
                    "required": true,
                    "properties":
                    {
                        "emailAdmin":{"type":"string"}
                    }              
                }
            ]
        },
        "/admin/profileclient":{
            "post":
            {
                "tags":["Admin Calls"],
                "summary": "Consult profile client",
                "description":"Call responsible for listing the customer profile",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                },
                {
                    "in": "body",
                    "name":"userData",
                    "description": false,
                    "required": true,
                    "properties":
                    {
                        "emailClient":{"type":"string"}
                    }              
                }
            ]
        },
        "/admin/editclientprofile":{
            "post":
            {
                "tags":["Admin Calls"],
                "summary": "Edit profile client",
                "description":"Call responsible for edit the customer profile",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                },
                {
                    "in": "body",
                    "name":"userData",
                    "description": false,
                    "required": true,
                    "properties":
                    {
                        "email":{"type":"string"}
                    }              
                }
            ]
        },
        "/admin/deleteclient/:emailclient":{
            "delete":
            {
                "tags":["Admin Calls"],
                "summary": "Delete client",
                "description":"Call responsible for delete the customer profile",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                }
            ]
        },
        "/admin/editacamera":{
            "post":
            {
                "tags":["Admin Calls"],
                "summary": "Edit camera client",
                "description":"Responsible route to edit the cameras within the ADMIN, CLIENT and WOWZA documents",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                },
                {
                    "in": "body",
                    "name":"userData",
                    "description": false,
                    "required": true,
                    "properties":
                    {
                        "emailClient":{"type":"string"},
                        "emailAdmin":{"type":"string"},
                        "descricao":{"type":"string"},
                        "RTSP":{"type":"string"},
                        "nomeCamera":{"type":"string"}
                    }              
                }
            ]
        },
        "/admin/deletecam":{
            "post":
            {
                "tags":["Admin Calls"],
                "summary": "Edit camera client",
                "description":"Route responsible for deleting customer's camera",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                },
                {
                    "in": "body",
                    "name":"userData",
                    "description": false,
                    "required": true,
                    "properties":
                    {
                        "emailClient":{"type":"string"},
                        "emailAdmin":{"type":"string"},
                        "nomeCamera":{"type":"string"}
                    }              
                }
            ]
        },
        "/admin/addgrupo":{
            "post":
            {
                "tags":["Admin Calls"],
                "summary": "Add group",
                "description":"Route responsible for creating a group to insert customers and cameras",
                "responses":
                {
                    "200":{"description":"authorized"},
                    "401":{"description":"not authorized"}
                }
            },
            "parameters":
            [
                {
                    "in": "header",
                    "name":"TOKEN_USER",
                    "description": "Bearer Token",
                    "required": true
                },
                {
                    "in": "body",
                    "name":"userData",
                    "description": false,
                    "required": true,
                    "properties":
                    {
                        "nomeGrupo":{"type":"string"},
                        "cameras":{"type":"string"},
                        "emailAdmin":{"type":"string"},
                        "emailClients":{"type":"string"}
                    }              
                }
            ]
        },
...
    }
}